+++ test/spec/fixtures/repos/spring-3-gradle-8-groovy/src/test/java/com/example/spring_3_gradle_groovy/DemoApplicationTests.java
@@ -1,0 +1,13 @@
+package com.example.spring_3_gradle_groovy;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DemoApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}